When testing for whether or not we needed <?xml-stylesheet type="text/css" href="../css/teibp.css"?>, I discovered that the display of code in egXML was not working in firefox. Upon investigating, I learned that the Mozilla developers refuse to support the disable-output-escaping attribute on <xsl:text>, which I was using to get the code to display.
As a workaround, I used some code from my regular tei2html stylesheets. I am importing a modified version of Evan Lenz's xml-to-string.xsl utility, which doesn't use any particular open source license but allows "Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met…" The conditions are just to retain this copyright statement and list of conditions.

The disadvantage of this approach is that it adds complexity. The advantage is that it works in Firefox, *and* it allows syntax highlighting, which is from code (xslt and css) that I added to the utility and to our stylesheets. 

Another disadvantage is that probably *very* few people will need this functionality. Most people are not wanting to display code in their TEI documents. 

Sooo… we should discuss this at our next meeting. 

I've made the questionable decision to update the live site, because without the update, everyone using firefox is going t have a broken demo, and the styles documentation will be broken. 

I'm surprised no one reported this problem.

git-svn-id: https://subversion.assembla.com/svn/tei_boilerplate@209 3bbcc0c9-51c9-4ab4-afdd-1b50544f6a2e