fascimile page viewer
Isolated problem with Chrome and Safari (i.e., WebKit) not jumping to
correct page image. The facs attribute had been set to a remote URL. If
we store images locally on same server as TEI Boilerplate, it works.
Need to find a solutin to this problem. For now I am taking easy way
out for demo adn storing images with TEI Boilerplate.