Fixed bug with tei title
Renamed teiHeader//title to tei-title. A hack because JavaScript was
doing weird things with &lt;title>, probably due to confusion with HTML
title. There is no TEI namespace in the TEI Boilerplate output because
JavaScript, or at least JQuery, cannot manipulate the TEI
elements/attributes if they are in the TEI namespace, so the TEI
namespace is stripped from the output. As far as I know, &lt;title>
elsewhere does not cause any problems, but we may need to extend this
to other occurrences of &lt;title> outside the Header.